(BSR)[API] refactor: move 'API Contremarque' route files to routes.public package
