+++ api/src/pcapi/routes/pro/public_api/__init__.py
@@ -2,8 +2,6 @@ from flask import Flask
 
 
 def install_routes(app: Flask) -> None:
-    # pylint: disable=unused-import
-    from . import bookings
     from .collective import install_routes as collective_install_routes
 
     collective_install_routes(app)
