+++ api/src/pcapi/routes/public/__init__.py
@@ -2,8 +2,10 @@ from flask import Flask
 
 
 def install_routes(app: Flask) -> None:
+    from . import booking_token
     from . import books_stocks
     from . import individual_offers
 
+    booking_token.install_routes(app)
     individual_offers.install_routes(app)
     books_stocks.install_routes(app)
