+++ api/src/pcapi/routes/public/booking_token/__init__.py
@@ -1,0 +1,9 @@
+from flask import Flask
+
+
+def install_routes(app: Flask) -> None:
+    from . import v1
+    from . import v2
+
+    v1.install_routes(app)
+    v2.install_routes(app)
