+++ api/src/pcapi/routes/public/booking_token/v1/__init__.py
@@ -1,0 +1,6 @@
+from flask import Flask
+
+
+def install_routes(app: Flask) -> None:
+    # pylint: disable=unused-import
+    from . import endpoints
