+++ api/tests/routes/public/booking_token/v1/patch_booking_by_token_test.py
