+++ api/tests/routes/public/booking_token/v2/patch_cancel_booking_by_token_test.py
